+++ src-tauri/src/lib.rs
@@ -77,6 +77,8 @@ pub fn run() {
         .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_fs::init())
+        .plugin(tauri_plugin_updater::Builder::new().build())
+        .plugin(tauri_plugin_process::init())
         .invoke_handler(tauri::generate_handler![load_data, save_data])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");